Fix bug where negative sleep could happen
The time_until_next_update value can be negative if the
time_spent_updating value is greater than the period value. This would
mean that the update and render process took longer than the desired
period, so the next iteration of the loop would need to start
immediately in order to keep up with the desired rate.